requirements.txt
absl-py==2.1.0
anyio==4.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.6.2
Brotli==1.0.9
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.0
executing==0.8.3
fastjsonschema==2.20.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.51.0
gast==0.4.0
google-auth==2.36.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.68.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.2
httpx==0.27.0
idna==3.7
importlib_metadata==8.5.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.15.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
json5==0.9.25
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
jupyter_client==8.6.0
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.0
jupyter_server==2.14.1
jupyter_server_terminals==0.4.4
jupyterlab==4.2.5
jupyterlab-pygments==0.1.2
jupyterlab_server==2.27.3
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.3
matplotlib==3.9.2
matplotlib-inline==0.1.6
mistune==2.0.4
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
nbclient==0.8.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.2
notebook_shim==0.2.3
numpy==1.26.4
oauthlib==3.2.2
opt_einsum==3.4.0
overrides==7.4.0
packaging==24.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
pillow==11.0.0
pip==24.3.1
platformdirs==3.10.0
ply==3.11
prometheus_client==0.21.0
prompt-toolkit==3.0.43
protobuf==3.19.6
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
Pygments==2.15.1
pyparsing==3.2.0
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
pywin32==305.1
pywinpty==2.0.10
PyYAML==6.0.2
pyzmq==25.1.2
referencing==0.30.2
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
rsa==4.9
Send2Trash==1.8.2
setuptools==75.1.0
sip==6.7.12
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.2.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.11.0
unicodedata2==15.1.0
urllib3==2.2.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
wheel==0.44.0
win-inet-pton==1.1.0
wrapt==1.17.0
zipp==3.21.0
